cmake: Fix warning finding libusb
UHD has a custom file to find libusb. This fixes a warning coming from
that file caused by the fact that we're looking for a package called
LIBUSB, but the file was called FindUSB1 (i.e., we're expecting
a package name of USB1).

Common CMake calls were also moved to lowercase for CMake coding
guidelines consistency.