chdr: Assert Ctrl Packet Size Before Serializing
This commit alters the ctrl_payload#serialize function to assert the
serialized length of the packet fits inside the buffer before writing.
Originally, the function writes the data and then asserts that the final
pointer position isn't further than the end of the buffer. While this
works, if it were to fail, that means we have already caused undefined
behavior by writing to some random memory location, which isn't good
practice.

This commit is possible now that the get_length method exists.

Signed-off-by: Samuel O'Brien <sam.obrien@ni.com>