example: Check for failure in tx_samples_from_file
I was using this example for testing with the simulator. If there is a
flow control failure, the original example would just silently finish,
outputing the message "Done!" (Not even printing a timeout message).

This commit asserts that the number of samples sent is equal to the
number of samples provided.

Signed-off-by: Samuel O'Brien <sam.obrien@ni.com>