rfnoc: Fix remote stream buffer format
When configuring remote streams, we were setting the format at the
source stream endpoint, but not at the destination stream endpoint.
Therefore, the destination used the default or whatever it was set to
during a previous run. This change sets the format at the destination
stream to match the format of the source stream.