sim: Fix CHDR header stringification
Prior to Boost 1.66, boost::format() did not support the %b format
specifier, yet the minimum version of Boost required to build UHD is
Boost 1.58 (as specified in the CMakeLists.txt file).

Rather than force an upgrade of Boost on everyone, this commit replaces
the %b format specifiers with %c and provides 'Y' or 'N' values based on
the Booleans in the CHDR header being printed (EOV and EOB).