endianness: Replace Boost macros with custom ones
This commit is a variation of a similar commit on UHD-3.15.LTS branch,
adapted to UHD 3.9.  The message below is copied from said commit.

Boost changed the macros for endianness identification in 1.69, and the
deprecation warning is a pretty noisy one during compilation. This
abstracts away the Boost macro, so we have two UHD macros,
UHD_BIG_ENDIAN and UHD_LITTLE_ENDIAN. They indicate big and little
endian byte order.