dboard_iface: Modify sleep() function
The algorithm for the sleep() function is changed to first increment the
command time if the command time is set.  If the command time is not
set, it just performs a sleep on the host.  The intention is to make a
best effort to create the requested delay on the device.

Signed-off-by: Michael West <michael.west@ettus.com>