ci: Build uhd with Pipelines
This commit adds the infrastructure to build uhd on both Linux and
Windows using Pipelines. Using the docker images generated by the
build docker pipeline, this runs an AzDO container job to build uhd
artifacts. It then stores the artifacts to be used in later pipelines or
stages. This change supports make and ninja on Linux and msbuild
with VS2017/VS2019 x86/x64 on Windows.

Signed-off-by: Steven Koo <steven.koo@ni.com>