ci: support for x3xx devtest on rhombus
Rhombus is a test suite that is intended run devtests for multiple
devices in parallel. This commit adds support for x300 and x310
support. This uses redis to mutex access to the hardware between
Azure Pipeline agents. This also updates the fpga using Vivado
over ssh to the host machine.

Signed-off-by: Steven Koo <steven.koo@ni.com>