uhd: Update all headers for setters on multi_usrp re coerce/throw
Our APIs are not consistent when it comes to handling invalid settings.
Some setting (like antenna, LO name, ...) will trigger an exception when
invalid. Other settings (gain, frequency, clock rate) will get coerced
to a valid value.

This behaviour does make sense for the most part (it is more intuitive
that 81 dB gets coerced to 80 dB if that's the maximum, but coercing an
invalid antenna value like "RX1" has no clear alternative). And in any
case, this is the behaviour that UHD has always had.

In this commit, all Doxygen headers in multi_usrp are updated to exactly
describe their behaviour (coerce or throw).