multi_usrp_rfnoc: Fix get_device()->get_tree()
Property tree access via the device tree was resulting in a segmentation
fault because the device::_tree member was not properly initialized for
the redirector_device class.  The redirector_device also had a
get_tree() method that was not actually called because the get_tree()
method in the parent class was not declared virtual and all calls to
get_tree() were done on the parent class (i.e. get_device()->get_tree().

- Added initializer for device::_tree member in redirector_device class
  constructor
- Removed redirector_device::get_tree() method

Signed-off-by: michael-west <michael.west@ettus.com>